Added error handling for all controllers and updated the error screen

+++ src/main/java/com/example/billTracker/controllers/CompanyController.java
@@ -11,10 +11,13 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.servlet.view.RedirectView;
 
+import com.example.billTracker.dto.BillDto;
 import com.example.billTracker.dto.CompanyDto;
+import com.example.billTracker.helper.ErrorMessage;
+import com.example.billTracker.repositories.BillRepository;
 import com.example.billTracker.repositories.CompanyRepository;
+
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -24,6 +27,9 @@ public class CompanyController{
 
 	@Autowired
 	CompanyRepository companyRepository;
+	
+	@Autowired
+	BillRepository billRepository;
 
 	private CompanyDto editCompany;
 
@@ -34,9 +40,13 @@ public class CompanyController{
 
 		model.addAttribute("companies", companies);
 
-		if(companies.isEmpty())
-			return "error";
-		return "company/all";
+		if(!companies.isEmpty()) return "company/all";
+			
+		ErrorMessage errorMessage = new ErrorMessage("Could not find any companies saved.");
+		
+		model.addAttribute("errorObject", errorMessage);
+		
+		return "error";
 	}
 
 	@PostMapping("/editCompany")
@@ -50,25 +60,58 @@ public class CompanyController{
 
 		if(company.isPresent())
 			return "company/edit";
+		
+		ErrorMessage errorMessage = new ErrorMessage("Could not find the selected company.");
+		
+		model.addAttribute("errorObject", errorMessage);
+		
 		return "error";
 	}
 
 	@PostMapping("/edit")
-	public RedirectView editCompany(@ModelAttribute("company") CompanyDto company){
+	public String editCompany(@ModelAttribute("company") CompanyDto company){
 
 		companyRepository.updateCompany(company.getCompanyName(), editCompany.getCompanyId());
 
 		editCompany = null;
 
-		return new RedirectView("http://localhost:8080/company/getAll");
+		return "redirect:/company/getAll";
 	}
 
 	@PostMapping("/delete")
-	public RedirectView deleteCompany(@RequestParam("companyId") String companyId, Model model){
+	public String deleteCompany(@RequestParam("companyId") String companyId, Model model){
 
+		Optional<CompanyDto> company = companyRepository.findById(Integer.parseInt(companyId));
+		
+		if(company.get() != null && company.get().isParentCompany()) {
+			ErrorMessage errorMessage = new ErrorMessage("Could not delete a parent company.");
+			
+			model.addAttribute("errorObject", errorMessage);
+			
+			return "error";
+		}else if(company.get() == null){
+			ErrorMessage errorMessage = new ErrorMessage("Could not find the selected company.");
+			
+			model.addAttribute("errorObject", errorMessage);
+			
+			return "error";
+		}
+		
+		int id = company.get().getCompanyId();
+		
+		List<BillDto> bills = billRepository.findBillsByPayingCompanyCompanyIdIsOrReceivingCompanyCompanyIdIs(id, id);
+		
+		if(!bills.isEmpty()) {
+			ErrorMessage errorMessage = new ErrorMessage("Could not delete the selected Company because it is saved in a Bill.");
+			
+			model.addAttribute("errorObject", errorMessage);
+			
+			return "error";
+		}
+		
 		companyRepository.deleteById(Integer.parseInt(companyId));
 
-		return new RedirectView("http://localhost:8080/company/getAll");
+		return "redirect:/company/getAll";
 	}
 
 	@GetMapping("/add")
@@ -82,17 +125,11 @@ public class CompanyController{
 	}
 
 	@PostMapping("/add")
-	public RedirectView companyAdded(CompanyDto company){
+	public String companyAdded(CompanyDto company){
 
 		companyRepository.save(company);
 
-		return new RedirectView("http://localhost:8080/company/getAll");
-	}
-
-	@GetMapping("/findById")
-	public String findById(){
-
-		return "";
+		return "redirect:/company/getAll";
 	}
 
 }
