+++ src/main/java/com/example/billTracker/controllers/EmployeeController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.elasticsearch.core.aggregation.impl.AggregatedPageImpl;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -15,11 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.view.RedirectView;
 
-import com.example.billTracker.config.EmployeeListBean;
-import com.example.billTracker.dto.BillDto;
-import com.example.billTracker.dto.CountryDto;
 import com.example.billTracker.dto.EmployeeDto;
 import com.example.billTracker.dto.SalaryDto;
+import com.example.billTracker.helper.ErrorMessage;
 import com.example.billTracker.repositories.EmployeeRepository;
 import com.example.billTracker.repositories.SalaryRepository;
 
@@ -35,23 +35,21 @@ public class EmployeeController{
 	
 	EmployeeDto editEmployee;
 	
-	@GetMapping("/bla")
-	public String bla() {
-		System.out.println(employeeRepository.findTopByOrderByEmployeeIdDesc().getEmployeeId());
-		
-		return "home";
-	}
-	
 	@GetMapping("/getAll")
 	public String getAll(Model model){
 
-		AggregatedPageImpl<EmployeeDto> employees = (AggregatedPageImpl<EmployeeDto>) employeeRepository.findAll();
+		PageImpl<EmployeeDto> employees = (PageImpl<EmployeeDto>) employeeRepository.findAll();
 
 		model.addAttribute("employees", employees);
 
-		if(employees.isEmpty())
-			return "error";
-		return "employee/getAll";
+		if(!employees.isEmpty()) return "employee/getAll";
+			
+		System.out.println("Ovde");
+		ErrorMessage errorMessage = new ErrorMessage("Could not find any employees.");
+		
+		model.addAttribute("errorObject", errorMessage);
+		
+		return "error";
 	}
 
 	@GetMapping("/add")
@@ -65,26 +63,35 @@ public class EmployeeController{
 	}
 
 	@PostMapping("/add")
-	public RedirectView countryAdded(EmployeeDto employee){
+	public String countryAdded(EmployeeDto employee){
 		 
 		EmployeeDto e = employeeRepository.findTopByOrderByEmployeeIdDesc();
-		int id = e.getEmployeeId() + 1;
+		
+		int id; 
+		
+		if(e == null) {
+			id = 1;
+		}else {
+			id = e.getEmployeeId() + 1;
+		}
 
 		employee.setEmployeeId(id);
+		
 		employeeRepository.save(employee);
 
-		return new RedirectView("http://localhost:8080/employee/getAll");
+		return "redirect:/employee/getAll";
 	}
 
 	List<EmployeeDto> getAllWithoutSalary(){
 
 		List<EmployeeDto> employeesWithoutSalaries = new LinkedList<>();
 
-		AggregatedPageImpl<EmployeeDto> employees = (AggregatedPageImpl<EmployeeDto>) employeeRepository.findAll();
+		PageImpl<EmployeeDto> employees = (PageImpl<EmployeeDto>) employeeRepository.findAll();
 
-		AggregatedPageImpl<SalaryDto> salaries = (AggregatedPageImpl<SalaryDto>) salaryRepository.findAll();
+		PageImpl<SalaryDto> salaries = (PageImpl<SalaryDto>) salaryRepository.findAll();
 
 		for(EmployeeDto e : employees){
+			if(salaries.isEmpty()) employeesWithoutSalaries.add(e);
 			for(SalaryDto s : salaries){
 				if(s.getEmployee().getEmployeeId() == e.getEmployeeId()){
 					break;
@@ -100,11 +107,23 @@ public class EmployeeController{
 	}
 	
 	@PostMapping("/delete")
-	public RedirectView deleteEmployee(@RequestParam("employeeId") String employeeId, Model model) {
+	public String deleteEmployee(@RequestParam("employeeId") String employeeId, Model model) {
 
-	    employeeRepository.deleteById(Integer.parseInt(employeeId));
+		SalaryDto salary = salaryRepository.findByEmployeeEmployeeId(Integer.parseInt(employeeId));
+		
+		if(salary == null) {			
+			employeeRepository.deleteById(Integer.parseInt(employeeId));
+
+			return "redirect:/employee/getAll";
+		}
+		
+		
+		ErrorMessage errorMessage = new ErrorMessage("The selected Employee could not be deleted because it is contained in a Salary.");
+		
+		model.addAttribute("errorObject", errorMessage);
+		
+	    return "error";
       
-	    return new RedirectView("http://localhost:8080/employee/getAll");
 	}
 	
 	@PostMapping("/editEmployee")
@@ -117,11 +136,16 @@ public class EmployeeController{
 	    model.addAttribute("employee", employee.get());
 	        
 	    if(employee.isPresent()) return "employee/edit";
+	    
+	    ErrorMessage errorMessage = new ErrorMessage("The selected employee could not be found.");
+		
+		model.addAttribute("errorObject", errorMessage);
+		
 	    return "error";
 	}
 	
 	@PostMapping("/edit")
-	public RedirectView editEmployee(@ModelAttribute("employee") EmployeeDto employee) {
+	public String editEmployee(@ModelAttribute("employee") EmployeeDto employee) {
 						
 		employee.setEmployeeId(editEmployee.getEmployeeId());
 		
@@ -129,6 +153,6 @@ public class EmployeeController{
 		
 		editEmployee = null;
 		
-		return new RedirectView("http://localhost:8080/employee/getAll");
+		return "redirect:/employee/getAll";
 	}
 }
